Revert "Refactor to remove duplicated code after Presentation API is re-implemented on Android"
This commit broke the Windows build:
    c:\b\b\slave\crosswalk_windows\build\src\xwalk\runtime\browser\xwalk_presentation_service_delegate_win.cc(76): error C2653: 'PresentationSessionWin' : is not a class or namespace name
    c:\b\b\slave\crosswalk_windows\build\src\xwalk\runtime\browser\xwalk_presentation_service_delegate_win.cc(76): error C3861: 'Create': identifier not found
    c:\b\b\slave\crosswalk_windows\build\src\xwalk\runtime\browser\xwalk_presentation_service_helper_win.cc(105) : error C2664: 'void xwalk::Runtime::set_observer(xwalk::Runtime::Observer *)' : cannot convert argument 1 from 'xwalk::PresentationSession *' to 'xwalk::Runtime::Observer *'
        Types pointed to are unrelated; conversion requires reinterpret_cast, C-style cast or function-style cast
    c:\b\b\slave\crosswalk_windows\build\src\xwalk\runtime\browser\xwalk_presentation_service_helper_win.cc(106) : error C2039: 'runtimes_' : is not a member of 'xwalk::PresentationSession'
        c:\b\b\slave\crosswalk_windows\build\src\xwalk\runtime\browser\xwalk_presentation_service_helper.h(141) : see declaration of 'xwalk::PresentationSession'
    c:\b\b\slave\crosswalk_windows\build\src\xwalk\runtime\browser\xwalk_presentation_service_helper_win.cc(106) : error C2228: left of '.push_back' must have class/struct/union

This reverts commit 4c7b5b5e66cb170c19063f4bd7de66602ed3a0ff.

BUG=XWALK-6262
