+++ xwalk.gyp
@@ -287,18 +287,10 @@
         'runtime/browser/xwalk_platform_notification_service.h',
         'runtime/browser/xwalk_pref_store.cc',
         'runtime/browser/xwalk_pref_store.h',
-        'runtime/browser/xwalk_presentation_service_delegate.cc',
-        'runtime/browser/xwalk_presentation_service_delegate.h',
         'runtime/browser/xwalk_presentation_service_delegate_android.cc',
         'runtime/browser/xwalk_presentation_service_delegate_android.h',
         'runtime/browser/xwalk_presentation_service_delegate_win.cc',
         'runtime/browser/xwalk_presentation_service_delegate_win.h',
-        'runtime/browser/xwalk_presentation_service_helper.cc',
-        'runtime/browser/xwalk_presentation_service_helper.h',
-        'runtime/browser/xwalk_presentation_service_helper_android.cc',
-        'runtime/browser/xwalk_presentation_service_helper_android.h',
-        'runtime/browser/xwalk_presentation_service_helper_win.cc',
-        'runtime/browser/xwalk_presentation_service_helper_win.h',
         'runtime/browser/xwalk_render_message_filter.cc',
         'runtime/browser/xwalk_render_message_filter.h',
         'runtime/browser/xwalk_runner.cc',
